Added Custom Hook to get screen-size for responsiveness
