+++ src/products-view/EditProductDetails.jsx
@@ -1,0 +1,9 @@
+import React from 'react'
+
+const EditProductDetails = () => {
+  return (
+    <div>EditProductDetails</div>
+  )
+}
+
+export default EditProductDetails
